+++ packages/progress-bar/example/index.tsx
@@ -1,2 +1,0 @@
-import "@fremtind/jkl-core/core.min.css";
-import "../progress-bar.scss";