+++ packages/select-react/documentation/Example.tsx
@@ -1,4 +1,4 @@
-import React, { useState, ChangeEvent } from "react";
+import React, { FocusEvent, useState, ChangeEvent } from "react";
 import { ExampleComponentProps } from "@fremtind/jkl-portal-components";
 import { Select, NativeSelect } from "../src";
 import { LabelVariant } from "@fremtind/jkl-core";
@@ -11,14 +11,23 @@ export const Example = ({ boolValues, choiceValues }: ExampleComponentProps) => 
     const values = [
         { value: "firstvalue", label: "Value 1" },
         { value: "secondvalue", label: "Value 2" },
+        { value: "thirdvalue", label: "Value 3" },
+        { value: "fourthvalue", label: "Value 4" },
     ];
     const [value, setValue] = useState<string>();
-    const universalSetValue = (input: string | ChangeEvent<HTMLSelectElement>) => {
+    const universalSetValue = (input: string | ChangeEvent<HTMLSelectElement> | undefined) => {
         if (typeof input === "string") {
             setValue(input);
-        } else {
+        } else if (input) {
             setValue(input.target.value);
         }
+    };
+
+    const onFocus = (input: string | FocusEvent<HTMLSelectElement> | undefined) => {
+        console.log("Focus: ", input);
+    };
+    const onBlur = (input: string | FocusEvent<HTMLSelectElement> | undefined) => {
+        console.log("Blur: ", input);
     };
 
     const errorLabel = boolValues && boolValues["Med feil"] ? "Beskrivende feilmelding" : undefined;
@@ -35,6 +44,8 @@ export const Example = ({ boolValues, choiceValues }: ExampleComponentProps) => 
             helpLabel={helpLabel}
             errorLabel={errorLabel}
             onChange={universalSetValue}
+            onBlur={onBlur}
+            onFocus={onFocus}
         />
     );
 };