+++ packages/text-input-react/src/TextArea.tsx
@@ -1,10 +1,9 @@
-import React, { forwardRef, FocusEvent, useRef, useState, RefObject } from "react";
+import React, { forwardRef, FocusEvent, useRef, useState, useEffect, RefObject } from "react";
 import classNames from "classnames";
 import nanoid from "nanoid";
 import { Label, SupportLabel } from "@fremtind/jkl-typography-react";
 import { LabelVariant } from "@fremtind/jkl-core";
 import { BaseProps } from "./BaseInputField";
-import { useExpandingTextArea } from "./useExpandingTextArea";
 
 interface Props extends BaseProps {
     label: string;
@@ -43,10 +42,38 @@ export const TextArea = forwardRef<HTMLTextAreaElement, Props>(
         const [supportId] = useState(`jkl-support-label-${nanoid(8)}`);
 
         const [textAreaFocused, setTextAreaFocused] = useState(false);
+        const [baseScrollHeight, setBaseScrollHeight] = useState(0);
+        const [currentRows, setCurrentRows] = useState(1);
         const internalRef = useRef<HTMLTextAreaElement>(null);
         const textAreaRef = (ref as RefObject<HTMLTextAreaElement>) || internalRef;
 
-        const { textAreaRows } = useExpandingTextArea(textAreaRef, rows, restProps.value, autoExpand, textAreaFocused);
+        useEffect(() => {
+            const textAreaElement = textAreaRef.current;
+            if (autoExpand && textAreaElement) {
+                const savedValue = textAreaElement.value;
+                const savedPlaceholder = textAreaElement.placeholder;
+                // BaseScrollHeight must be calculated from an empty textarea and empty placeholder.
+                textAreaElement.value = "";
+                textAreaElement.placeholder = "";
+                setBaseScrollHeight(textAreaElement.scrollHeight);
+                textAreaElement.value = savedValue;
+                textAreaElement.placeholder = savedPlaceholder;
+            }
+        }, [textAreaRef, autoExpand]);
+
+        useEffect(() => {
+            const textAreaElement = textAreaRef.current;
+            const minimumRows = rows;
+
+            if (textAreaElement) {
+                const calculatedRows = calculateRows(textAreaElement, baseScrollHeight);
+                if (textAreaFocused || restProps.value) {
+                    setCurrentRows(Math.max(minimumRows, calculatedRows));
+                } else {
+                    setCurrentRows(calculatedRows);
+                }
+            }
+        }, [textAreaRef, restProps.value, textAreaFocused, baseScrollHeight, rows]);
 
         function onFocus(e: FocusEvent<HTMLTextAreaElement>) {
             setTextAreaFocused(true);
@@ -58,21 +85,34 @@ export const TextArea = forwardRef<HTMLTextAreaElement, Props>(
             restProps.onBlur && restProps.onBlur(e);
         }
 
+        function calculateRows(textAreaElement: HTMLTextAreaElement, baseScrollHeight: number) {
+            const lineHeightWithPx = window.getComputedStyle(textAreaElement).lineHeight;
+            const lineHeight = parseInt(lineHeightWithPx.replace("px", ""));
+            const savedRows = textAreaElement.rows;
+            // We need to set rows to 1 to shrink the textarea when removing characters.
+            textAreaElement.rows = 1;
+            const calculatedRows = Math.ceil((textAreaElement.scrollHeight - baseScrollHeight) / lineHeight) + 1;
+            textAreaElement.rows = savedRows;
+            return calculatedRows;
+        }
+
         return (
             <div data-testid="jkl-text-area" className={componentClassName}>
                 <Label standAlone htmlFor={uid} variant={variant} forceCompact={forceCompact}>
                     {label}
                 </Label>
                 <textarea
+                    id={uid}
+                    ref={textAreaRef}
+                    className={`jkl-text-input__input jkl-text-input__input--${rows}-rows`}
                     onFocus={onFocus}
                     onBlur={onBlur}
-                    ref={textAreaRef}
                     aria-invalid={!!errorLabel}
-                    className={`jkl-text-input__input jkl-text-input__input--${rows}-rows`}
-                    id={uid}
-                    rows={textAreaRows}
-                    placeholder={placeholder}
                     aria-describedby={supportId}
+                    placeholder={placeholder}
+                    rows={autoExpand ? currentRows : undefined}
+                    // Must set overflowX hidden for Firefox https://stackoverflow.com/a/22700700
+                    style={autoExpand ? { height: "auto", overflowX: "hidden" } : undefined}
                     {...restProps}
                 />
                 <SupportLabel
@@ -86,4 +126,5 @@ export const TextArea = forwardRef<HTMLTextAreaElement, Props>(
         );
     },
 );
+
 TextArea.displayName = "TextArea";