+++ portal/src/components/Typography/Typography.tsx
@@ -1,46 +1,81 @@
-import React, { HTMLAttributes } from "react";
+import React, { ReactElement } from "react";
+import { Prism as SyntaxHighlighter } from "react-syntax-highlighter";
+import { prism } from "react-syntax-highlighter/dist/esm/styles/prism";
 
-type HeadingProps = HTMLAttributes<HTMLHeadingElement>;
-type ParagraphProps = HTMLAttributes<HTMLParagraphElement>;
-
-export const PageTitle = ({ children, ...rest }: HeadingProps) => (
+export const PageTitle: React.FC = ({ children, ...rest }) => (
     <h1 className="jkl-title-large jkl-portal-page-title" {...rest}>
         {children}
     </h1>
 );
 
-export const HeadingLarge = ({ children, ...rest }: HeadingProps) => (
+export const HeadingLarge: React.FC = ({ children, ...rest }) => (
     <h2 className="jkl-heading-large jkl-portal-heading-large" {...rest}>
         {children}
     </h2>
 );
 
-export const HeadingMedium = ({ children, ...rest }: HeadingProps) => (
+export const HeadingMedium: React.FC = ({ children, ...rest }) => (
     <h3 className="jkl-heading-medium jkl-portal-heading-medium" {...rest}>
         {children}
     </h3>
 );
 
-export const HeadingSmall = ({ children, ...rest }: HeadingProps) => (
+export const HeadingSmall: React.FC = ({ children, ...rest }) => (
     <h4 className="jkl-heading-small jkl-portal-heading-small" {...rest}>
         {children}
     </h4>
 );
 
-export const HeadingXS = ({ children, ...rest }: HeadingProps) => (
+export const HeadingXS: React.FC = ({ children, ...rest }) => (
     <h5 className="jkl-micro jkl-portal-heading-xs" {...rest}>
         {children}
     </h5>
 );
 
-export const ArticleLead = ({ children, ...rest }: ParagraphProps) => (
+export const ArticleLead: React.FC = ({ children, ...rest }) => (
     <p className="jkl-lead jkl-portal-article-lead" {...rest}>
         {children}
     </p>
 );
 
-export const Paragraph = ({ children, ...rest }: ParagraphProps) => (
+export const Paragraph: React.FC = ({ children, ...rest }) => (
     <p className="jkl-body jkl-portal-paragraph" {...rest}>
         {children}
     </p>
 );
+
+export const InlineCode: React.FC = ({ children, ...rest }) => (
+    <code className="jkl-portal-inline-code" {...rest}>
+        {children}
+    </code>
+);
+
+export const CodeBlock: React.FC = ({ children, ...rest }) => {
+    const child: ReactElement = React.Children.only(children) as ReactElement;
+    if (!child.props) {
+        // if there is more than one child, or it is text return a normal pre
+        return <pre {...rest}>{children}</pre>;
+    }
+    const language = child.props.className?.replace("language-", "");
+    const style = {
+        ...prism,
+        'pre[class*="language-"]': {
+            lineHeight: "1.4",
+        },
+        operator: {
+            ...prism.operator,
+            background: "transparent",
+        },
+    };
+    return (
+        <SyntaxHighlighter
+            className="jkl-portal-code-block"
+            style={style}
+            codeTagProps={{ style: {}, className: "jkl-portal-code-block__code" }}
+            language={language}
+            data-language={language}
+        >
+            {child.props.children}
+        </SyntaxHighlighter>
+    );
+};