+++ packages/logo-react/CHANGELOG.md
@@ -3,6 +3,37 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.3](https://github.com/fremtind/jokul/compare/@fremtind/jkl-logo-react@3.0.2...@fremtind/jkl-logo-react@3.0.3) (2020-04-20)
+
+**Note:** Version bump only for package @fremtind/jkl-logo-react
+
+
+
+
+
+## [3.0.2](https://github.com/fremtind/jokul/compare/@fremtind/jkl-logo-react@3.0.1...@fremtind/jkl-logo-react@3.0.2) (2020-03-30)
+
+
+### Bug Fixes
+
+* **logo:** add transforms for logo symbol support ([fb73270](https://github.com/fremtind/jokul/commit/fb7327055ee15c298f8e4a02b0e2149163dfae78))
+* **logo-react:** use .logo-example on LogoExample ([e067d18](https://github.com/fremtind/jokul/commit/e067d18f385724ad34ab71f7f94cdeafbfb527a9))
+
+
+
+
+
+## [3.0.1](https://github.com/fremtind/jokul/compare/@fremtind/jkl-logo-react@3.0.0...@fremtind/jkl-logo-react@3.0.1) (2020-03-25)
+
+
+### Bug Fixes
+
+* move browserslist declaration to root package.json ([51c790e](https://github.com/fremtind/jokul/commit/51c790ea79ca3d667871380c6bfbe85a5738920b)), closes [#862](https://github.com/fremtind/jokul/issues/862)
+
+
+
+
+
 # [3.0.0](https://github.com/fremtind/jokul/compare/@fremtind/jkl-logo-react@2.1.1...@fremtind/jkl-logo-react@3.0.0) (2020-03-16)