+++ packages/typography-react/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.2.0](https://github.com/fremtind/jokul/compare/@fremtind/jkl-typography-react@2.1.3...@fremtind/jkl-typography-react@2.2.0) (2020-04-20)
+
+
+### Features
+
+* **links:** add external prop and style to links ([f1a47f0](https://github.com/fremtind/jokul/commit/f1a47f09cc7ed5c8a1c5ac19de8b9b8ed78ffac9)), closes [#905](https://github.com/fremtind/jokul/issues/905)
+
+
+
+
+
+## [2.1.3](https://github.com/fremtind/jokul/compare/@fremtind/jkl-typography-react@2.1.2...@fremtind/jkl-typography-react@2.1.3) (2020-03-25)
+
+
+### Bug Fixes
+
+* move browserslist declaration to root package.json ([51c790e](https://github.com/fremtind/jokul/commit/51c790ea79ca3d667871380c6bfbe85a5738920b)), closes [#862](https://github.com/fremtind/jokul/issues/862)
+
+
+
+
+
 ## [2.1.2](https://github.com/fremtind/jokul/compare/@fremtind/jkl-typography-react@2.1.1...@fremtind/jkl-typography-react@2.1.2) (2020-03-16)
 
 **Note:** Version bump only for package @fremtind/jkl-typography-react